So, this is an attempt at writing a game for the Github Game Jam "Game Off 2020".
I decided I will use the commit messages both for the usual tracking of changes as well as for keeping kind of a "diary" of my progress. Although the game jam will be going on through all of November, the limited amount of time I have available will probably mean that I won't finish the game, but whatevs!

So, for now, I'm just checking in a crude initial setup. I don't know what kind of game this is going to be but I've decided I'm going to do it in Javascript, even though I have very little experience with it, let alone with WebGL. So, even if this will never amount to a finished game, my hope is that at least I might get a little bit of experience out of it.

The theme of the game jam is "Moonshot", so I figure a good deal of the submitted games will be sci-fi based, and then some more or less clever/punny interpretation of the term for something completely different. I think I will stick to the sci-fi realm.

Let's keep our fingers crossed.